Updated PDB Parser with better code
